refactor: Add `BetterDecorationItem` class ('cause Mojang was too lazy)
Mojang's default decoration item class has a fatal flaw in terms of
extensibility: it doesn't automatically spawn the entity, because it
doesn't know what that entity's type is.

This `BetterDecorationItem` class expects the superclass to implement
`getEntityProvider`, which should be a lambda / method call that returns
the entity type. This allows the `WaveformEntity` class to be a bit
cleaner, and not require code that Mojang should've written in the first
place, instead of hardcoded if statements.